add loadservice

+++ Data/Entities/FeeInformation.cs
@@ -10,9 +10,9 @@ namespace Data.Entities
         public Loan Loan { get; set; }
         public int Number { get; set; }
         public DateTimeOffset Date { get; set; }
-        public double TotalFee { get; set; }
-        public double CurrentAmount { get; set; }
-        public double CapitalPayment{get;set;}
+        public decimal TotalFee { get; set; }
+        public decimal CurrentAmount { get; set; }
+        public decimal CapitalPayment{get;set;}
         public int Id {get; set; }
         public bool IsDeleted { get; set; }
         public DateTimeOffset CreationDate { get ; set; }
